+++ apps/content-analysis/src/utils/getMorphologyData.js
@@ -6,26 +6,28 @@ let morphologyData = null;
  * @returns {Object} The morphology data, or an empty object if not available.
  */
 function loadLocalMorphologyData() {
-	let data, dataDe, dataNL, dataES, dataFR, dataRU = {};
+	let data, dataDe, dataNL, dataES, dataFR, dataRU, dataIT = {};
 	try {
 		// Disabling global require to be able to fail.
 		// eslint-disable-next-line global-require
 		data = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-v3.json" );
 		// eslint-disable-next-line global-require
-		dataDe = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-de-v6.json" );
+		dataDe = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-de-v8.json" );
 		// eslint-disable-next-line global-require
-		dataNL = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-nl-v6.json" );
+		dataNL = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-nl-v8.json" );
 		// eslint-disable-next-line global-require
-		dataES = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-es-v6.json" );
+		dataES = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-es-v8.json" );
 		// eslint-disable-next-line global-require
-		dataFR = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-fr-v6.json" );
+		dataFR = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-fr-v8.json" );
 		// eslint-disable-next-line global-require
-		dataRU = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-ru-v6.json" );
+		dataRU = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-ru-v8.json" );
+		// eslint-disable-next-line global-require
+		dataIT = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-it-v8.json" );
 	} catch ( error ) {
 		// Falling back to empty data.
 	}
 
-	return merge( data, dataDe, dataNL, dataES, dataFR, dataRU );
+	return merge( data, dataDe, dataNL, dataES, dataFR, dataRU, dataIT );
 }
 
 /**