+++ packages/yoastseo/src/helpers/getStemForLanguage.js
@@ -4,6 +4,7 @@ import { determineStem as dutchDetermineStem } from "../morphology/dutch/determi
 import spanishDetermineStem from "../morphology/spanish/stem";
 import frenchDetermineStem from "../morphology/french/stem";
 import russianDetermineStem from "../morphology/russian/stem";
+import italianDetermineStem from "../morphology/italian/stem";
 
 /**
  * Collects all functions for determining a stem per language and returns this collection to a Researcher
@@ -18,5 +19,6 @@ export default function() {
 		es: spanishDetermineStem,
 		fr: frenchDetermineStem,
 		ru: russianDetermineStem,
+		it: italianDetermineStem,
 	};
 }