+++ packages/yoastseo/src/morphology/french/stem.js
@@ -1,6 +1,7 @@
 /* eslint-disable max-statements, complexity */
 import { checkIfWordEndingIsOnExceptionList } from "../morphoHelpers/exceptionListHelpers";
 import { applyAllReplacements } from "../morphoHelpers/regexHelpers";
+
 /*
  * MIT License
  *
@@ -285,11 +286,13 @@ const removeVerbSuffixesStartingWithI = function( word, originalWord, rvIndex, v
  * @param {string}  wordAfterStep1    The word after step 1 was done.
  * @param {number}  r2Index           The start index of R2.
  * @param {number}  rvIndex           The start index of RV.
- * @param {Object}  otherVerbSuffixes The French verbal suffixes that do not start with I.
+ * @param {Object}  morphologyData    The French morphology data.
  *
  * @returns {string} The word after other verb suffixes were removed.
  */
-const removeOtherVerbSuffixes = function( word, step2aDone, wordAfterStep1, r2Index, rvIndex, otherVerbSuffixes ) {
+const removeOtherVerbSuffixes = function( word, step2aDone, wordAfterStep1, r2Index, rvIndex, morphologyData ) {
+	const otherVerbSuffixes = morphologyData.regularStemmer.otherVerbSuffixes;
+
 	if ( step2aDone && wordAfterStep1 === word ) {
 		const suffixIons = new RegExp( otherVerbSuffixes[ 0 ] );
 		if ( word.search( suffixIons ) >= r2Index ) {
@@ -299,8 +302,18 @@ const removeOtherVerbSuffixes = function( word, step2aDone, wordAfterStep1, r2In
 		for ( let i = 1; i < otherVerbSuffixes.length; i++ ) {
 			const regex = new RegExp( otherVerbSuffixes[ i ] );
 			if ( word.search( regex ) >= rvIndex ) {
-				return word.replace( regex,  "" );
+				return word.replace( regex, "" );
 			}
+		}
+		// Check if a word ends in "ons" preceded by "i", if it is "ons" is not stemmed.
+		if ( word.endsWith( "ions" ) ) {
+			return word;
+		}
+
+		// Check if a word ends in "ons" preceded by other than "i" and stem it if it is in RV.
+		const verbSuffixOns = new RegExp( morphologyData.regularStemmer.verbSuffixOns );
+		if ( word.search( verbSuffixOns ) >= rvIndex ) {
+			word = word.replace( verbSuffixOns, "" );
 		}
 	}
 
@@ -375,9 +388,43 @@ const checkWordInFullFormExceptions = function( word, exceptions ) {
  * @returns {null|string}	The canonical stem if word was found on the list.
  */
 const canonicalizeStem = function( stemmedWord, stemsThatBelongToOneWord ) {
+	// Check the adjectives list.
 	for ( const paradigm of stemsThatBelongToOneWord.adjectives ) {
 		if ( paradigm.includes( stemmedWord ) ) {
 			return paradigm[ 0 ];
+		}
+	}
+	// Check the verbs list. The infinitive stem is always the canonical stem for verbs.
+	for ( const paradigm of stemsThatBelongToOneWord.verbs ) {
+		if ( paradigm.includes( stemmedWord ) ) {
+			return paradigm[ 0 ];
+		}
+	}
+};
+
+/**
+ * Checks whether the word is on the list of words which should be stemmed, even though the suffix would not be found in the
+ * required region. If the word is found on the list, the stem specified in that list is returned.
+ *
+ * @param {string}	word				The word to check.
+ * @param {Object}	shortWordsAndStems	The list to check.
+ *
+ * @returns {null|string} The stem or null if the word was not found on the list.
+ */
+const checkShortWordsExceptionList = function( word, shortWordsAndStems ) {
+	// First check whether the word is on the sub-list of words that cannot take an extra -s suffix.
+	for ( const wordStemPair of shortWordsAndStems.cannotTakeExtraSuffixS ) {
+		if ( wordStemPair[ 0 ] === word ) {
+			return wordStemPair[ 1 ];
+		}
+	}
+	// If the word was not found on the first sub-list, check the second sub-list of words that can take an extra -s suffix.
+	if ( word.endsWith( "s" ) ) {
+		word = word.slice( 0, -1 );
+	}
+	for ( const wordStemPair of shortWordsAndStems.canTakeExtraSuffixS ) {
+		if ( wordStemPair[ 0 ] === word ) {
+			return wordStemPair[ 1 ];
 		}
 	}
 };
@@ -394,6 +441,12 @@ export default function stem( word, morphologyData ) {
 	word = word.toLowerCase();
 	const originalWord = word;
 
+	// Check if the word is on an exception list of words that should be stemmed even though the suffix is not in the required region.
+	const wordAfterShortWordsCheck = checkShortWordsExceptionList( word, morphologyData.shortWordsAndStems );
+	if ( wordAfterShortWordsCheck ) {
+		return wordAfterShortWordsCheck;
+	}
+
 	// Check if the word is on an exception list for which all forms of a word and its stem are listed.
 	const ifException = checkWordInFullFormExceptions( word, morphologyData.exceptionStemsWithFullForms );
 	if ( ifException ) {
@@ -404,6 +457,35 @@ export default function stem( word, morphologyData ) {
 	if ( word.endsWith( "x" ) ) {
 		const pluralsWithXSuffix = morphologyData.pluralsWithXSuffix;
 		if ( pluralsWithXSuffix.includes( word ) ) {
+			return word.slice( 0, -1 );
+		}
+	}
+
+	// Check if the word is on the exception list of words for which -s should not be stemmed.
+	if ( word.endsWith( "s" ) ) {
+		const sShouldNotBeStemmed = morphologyData.sShouldNotBeStemmed;
+		if ( sShouldNotBeStemmed.includes( word ) ) {
+			return word;
+		}
+	}
+
+	// Check if the word is on the exception list of words for which -ent should not be stemmed.
+	const nonVerbsOnEnt = morphologyData.nonVerbsOnEnt;
+	if ( word.endsWith( "ent" ) ) {
+		if ( nonVerbsOnEnt.includes( word ) ) {
+			return word;
+		}
+	}
+	if ( word.endsWith( "ents" ) ) {
+		if ( nonVerbsOnEnt.includes( word.slice( 0, -1 ) ) ) {
+			return word.slice( 0, -1 );
+		}
+	}
+
+	// Check if word is on the exception list of nouns and adjectives for which the verb suffix -ons should not be stemmed.
+	const nonVerbsOnOns = morphologyData.nonVerbsOnOns;
+	if ( word.endsWith( "ons" ) ) {
+		if ( nonVerbsOnOns.includes( word ) ) {
 			return word.slice( 0, -1 );
 		}
 	}
@@ -442,14 +524,16 @@ export default function stem( word, morphologyData ) {
 	 * Step 2b:
 	 * Stem other verb suffixes
 	 */
-	word = removeOtherVerbSuffixes(
-		word,
-		step2aDone,
-		wordAfterStep1,
-		r2Index,
-		rvIndex,
-		morphologyData.regularStemmer.otherVerbSuffixes
-	);
+	if ( ! nonVerbsOnEnt.includes( word ) ) {
+		word = removeOtherVerbSuffixes(
+			word,
+			step2aDone,
+			wordAfterStep1,
+			r2Index,
+			rvIndex,
+			morphologyData
+		);
+	}
 
 	if ( originalWord === word.toLowerCase() ) {
 		/* Step 4: